+++ pineappl/tests/drell_yan_lo.rs
@@ -285,8 +285,6 @@ fn perform_grid_tests(
 
     let pdf_set = "NNPDF31_nlo_as_0118_luxqed";
 
-    assert!(lhapdf::available_pdf_sets().iter().any(|x| x == pdf_set));
-
     let pdf = Pdf::with_setname_and_member(pdf_set, 0)?;
     let mut xfx = |id, x, q2| pdf.xfx_q2(id, x, q2);
     let mut alphas = |_| 0.0;